v2/pytesting_os/requirements.txt
keystoneauth1==2.9.0
monotonic==1.1
msgpack-python==0.4.7
netaddr==0.7.18
netifaces==0.10.4
pyasn1==0.1.9
pyOpenSSL==16.0.0
python-cinderclient==1.8.0
python-glanceclient==2.1.0
python-heatclient==1.3.0
python-keystoneclient==3.2.0
python-neutronclient==4.2.0
python-novaclient==4.1.0
python-swiftclient==3.0.0
requests==2.10.0
requestsexceptions==1.1.3
six==1.10.0
nose==1.3.7
paramiko==2.0.2
python-dateutil==2.6.0